Fixes #1
Resolved an issue where users buying $EDU with other
token pairs were receiving WEDU instead of unwrapped
$EDU. The problem stemmed from a bug in the
`sendToken` implementation of the `TokenPayments`
library, which failed to unwrap WEDU before sending it
to the user. This update addresses and corrects the issue.

Important: For this upgrade to take effect, the Pairs and
Governance contracts need to be upgraded. Ensure that
the Pairs contract is upgraded first, as upgrading other
contracts before the Pairs may prevent the fix from
being applied correctly.